requirements.txt
Django==2.0.5
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.4
gunicorn==19.6.0
whitenoise==3.2.1